move the heapStats test into its own file so the gate does not also time the 10k-abort express fixture
